Init solver with Levenberg-Marquardt optimizer
This adds a limited number of constraints: two. (Point-point distance
and point-point-point angle.) Levenberg-Marquardt is implemented as the
solver; this method may be usable as a general solver, as it should be
able to handle under- and overconstrained systems. It will be
interesting to compare its performance (both in solving quality and
efficiency) to, e.g., (L-)BFGS as the general solver.

As the initial and experimental API, the constraint system is set up to
have separate elements and constraints. When an element is added, a
handle to it is returned that can be used to construct constraints. To
allow the user to choose various subsets to solve, elements and
constraints can be added to sets that the user can hand to the
`System::solve` method.

It is not yet very ergonomic to get elements in or out of the system, it
would be ideal to use Kurbo types, especially for output. For input,
it's somewhat harder, e.g., when inputting a `Line` you may want to reuse
existing `Point`s.