Add L-BFGS with Hager-Zhang line search (#3)
This implement L-BFGS with the algorithms as in the Numerical
Optimization textbook by Nocedal and Wright (2006). This is equivalent
to the [L-BFGS algorithm as described on
Wikipedia](https://en.wikipedia.org/w/index.php?title=Limited-memory_BFGS&oldid=1294232678),
modulo for what I believe to be a mistake in the calculation of γ. (A
recent diff on Wikipedia updated that to use the oldest values in the
stored history, whereas Nocedal and Wright use values of the previous
iteration.)

L-BFGS iterations find a direction to update the free variables towards,
and must be followed by a step size in that direction satisfying the
[Wolfe conditions](https://en.wikipedia.org/wiki/Wolfe_conditions). This
uses the line search method described in ["A new conjugate gradient
method with guaranteed descent and an efficient line search" by Hager
and Zhang (2005)](https://epubs.siam.org/doi/abs/10.1137/030601880),
which is suggested as line search algorithm by e.g. Nocedal and Wright,
and also implemented for L-BFGS specifically in e.g. TensorFlow.