Extract `calculate_residuals_and_jacobian` to `utils` (#4)
This functionality is needed for L-BFGS and Levenberg-Marquardt, and is
currently defined twice.

While we're here, this cleans up some unnecessary parameters passed to
the gradient calculation function in L-BFGS. Also improves naming of the
"free variable map" (though I'm not entirely happy with how variables,
constraints, free variables are represented at the moment).