Add point-line incidence constraint (#2)
Point-line incidence is a logical constraint, rather than a metric one.
It's somewhat similar to the vertex angle (`PointPointPointAngle`), with
an angle of 180 degrees, except for this incidence constraint it doesn't
matter which point is the vertex point. This could be called
`PointPointPointCollinearity` and use points rather than lines. We
should think about exactly how we want to name and represent these
elements and constraints.

This shows how in the current setup a `Line` just references two
existing `Point`s.