Output solved elements as concrete data (#7)
When getting data out of `System` you probably just want the concrete
values, and not to have to do the element handle-chasing yourself.

There is some messiness currently as the crate-private methods are on
the sealed inner trait `ElementInner`, and so the associated output type
is defined there as well. To allow using `Element::Output` in public
facing signatures then requires the associated type on `Element` to have
a type bound: `Element::Output: From<<Self as ElementInner>::Output>`,
enabling us to use the no-op blanket `From` implementation internally to
get things to type-check. Perhaps it's better to just move the methods
to the `Element` trait directly with a `#[doc(hidden)]`.