Call `create` on constraints to add them to the system
This flips the API: rather than constructing a constraint and then
adding it to the system, the constraint adds itself to the system and
returns a handle to itself.

On main, constraint structs must be constructed first and can only be
used for passing to `System::add_constraint` immediately after
construction, which is not very useful. This requires having duplicated
structs for constraints, as we want to store other information required
for performing the calculations inside the structs, not the element
handles used to construct them.

As there will potentially be many constraints, an API like
`System::{constrain_point_point_distance,
constrain_point_line_incidence_, ...` would probably grow too many
methods.

(In this PR, the `Constraint` trait is now mostly vestigial, but may be
useful in the future.)