Numerically find constraints causing systems to be overconstrained (#24)
This is based on the linearization in the system's current state
(through the Jacobian). Constraints that are linearly dependent on other
constraints (i.e., that can be written in terms of a linear combination
of other constraints) are designated as causing overconstrainedness.

Ideally, the system should first be partially solved for singular
constraints like parallelisms and incidences, generating a "witness" as
in "Geometric Constraint Solving: the Witness Configuration Method" by
Michelucci and Foufou (2006).