Add `Graph` to track structure of the system (#31)
This keeps track of connected components online (to know which parts of
the system are completely distinct and can trivially be solved and
analyzed separately). The graph also tracks the degrees of freedom of
the elements and valency of constraints. That information is not used
here yet, but will be soon.

The tracking is based on the element primitives. For example, a line
consists of exactly two point primitives, and nothing more. A circle
consists of a point primitive and a radius. That means a line-line
parallelism constraint, say, adds a hyperedge between four point
primitives.