Implement graph-based decomposition (#32)
This performs a degrees-of-freedom analysis on the graph of primitive
elements and constraints, finding a series of small steps that can be
solved to solve the entire system. The smaller the steps, the more
efficiently the system can be solved.

This is based on the Modified Frontier Algorithm described in
"Decomposition Plans for Geometric Constraint Problems, Part II: New
Algorithms" (2001) by C. M. Hoffman, A. Lomonosov and M. Sitharam.