Implement Levenberg-Marquardt solve with QR-decomposition (#48)
This is slower per computation and therefore slower in the
well-conditioned case, but it is numerically more stable as we don't
form the square of the Jacobian. In numerically ill-conditioned cases,
the lack of stability due to forming the square can lead to having to
perform many more iterations to converge, as well as failure to converge
at all.