Test a system that spans a large order of magnitude (#47)
It passes on top of https://github.com/endoli/fiksi/pull/48. Note that
it fails if not on top of that PR:
https://github.com/endoli/fiksi/commit/be4c37fa64c0007722a25ac0899fbd098f1aa947.

This is numerically tricky. The singular line-line parallelism
constraint residual (a cross product) will roughly be double the order
of magnitude of the point-point distance constraint residuals (a vector
norm).