Add point-point coincidence constraint
Fixes #62.

On top of #67, which introduces the required mechanism.

This constraint is quite different from the ones already in main, as it
has a valency of 2, i.e., it takes away two degrees of freedom. (Say one
of the two points constrained by a point-point coincidence is fixed,
then the other point has no degrees of freedom left. To satisfy the
constraint, it has to be at the exact same location.)

This is best modeled by introducing two residuals, one linear distance
along the x-axis and one along the y-axis, as this is numerically more
than a `PointPointDistance` constraint with `distance` parameter 0.