Refactor constraints to provide one or more expressions (#67)
This decouples constraints from residual calculation by introducing
expressions, allowing us to introduce constraints that provide more than
one residual. This is useful for having constraints with a valency
larger than one (like point-point coincidences, which have a valency of
2 in 2D).

Parts of the code working at the residual-level, like the optimizers,
can now continue working on a flat list of expressions, rather than
operating on a list of constraints with differing valency.