Add system-of-equation-based decomposition (#83)
Add a decomposition method for systems of equations to find sets of
expressions that must be solved together and the variables that are
considered free at each point.

This is more general than geometrical decomposition, and therefore
somewhat less powerful. It is however somewhat simpler and all current
tests pass when they're set to decompose. Decomposition is still off by
default for now.

It is possible this method can also be used within a system that has
been geometrically decomposed.