Implement a recursively assembling decomposer and solver
This continues with the existing (but dead) code doing the rigid cluster
analysis, but fixes finding of clusters that have to be rigidly
transformed at each step.

This is based on the Modified Frontier Algorithm described in
"Decomposition Plans for Geometric Constraint Problems, Part II: New
Algorithms" (2001) by C. M. Hoffman, A. Lomonosov and M. Sitharam.

A solver using this decomposition plan is implemented, that uses the
recursive assembly decomposition to solve clusters one-by-one, keeping
all elements that have occurred on previously-seen clusters rigid with
respect to those clusters' local coordinate systems.

Solve sets are removed for now, as that makes this implementation more
straightforward. As we have decided to no longer allow masking out
constraints, this should be replaced by being able to mark elements as
fixed or free.

This work is still experimental, as it is possible to break (probably
some bookkeeping issues), and sometimes systems fail to solve
(potentially also bookkeeping issues); but it should be mostly in the
right direction otherwise. Moderately complex systems get solved
correctly, but randomness in the decomposition (decomposition aren't
unique) can sometimes fail to solve such systems.

Next steps may be to align the decomposition planner more with the
structure described in the paper, which would allow for "design
decompositions" to be input (e.g., describing subsystems to be solved
first within the partial solving order, to assign "priorities"), as well
as potentially improve efficiency: the planner here still uses an
inefficient exhaustive search for dense subsystems.