Allow updating element values (#85)
This is the element counterpart of
https://github.com/endoli/fiksi/pull/78.

Closes https://github.com/endoli/fiksi/issues/59.

This doesn't do `Line` or `Circle`, as those reference `Point` and
`Length`. Perhaps we want to be able to update those directly too, but
then we'd probably want to be able to construct them directly from
values as well (creating and perhaps returning the necessary `Point`s
and `Length`s automatically).