Init Fiksi SVG renderer (#73)
This is a bit limited still. It allows rendering the geometric elements
in systems to an SVG. Elements can be given a color.

Run an example with the following.

```bash
cargo run -p fiksi_svg_tests
```

Note this doesn't do anything to visualize constraints themselves, which
would be nice to have, but also is somewhat difficult.