Add `Problem` to generalize optimizers, add `VariableMap` to map variables (#93)
For decomposition we need to be able to solve differently-structured
systems (e.g., solving for the pose of a rigid cluster, rather than that
cluster's primitive elements' variables).

The `Problem` and `VariableMap` introduced here are conceptually
orthogonal, but implementation-wise are easiest when done together. This
allows the numeric optimizers to take a generic `Problem` and a mutable
slice of just the free variables, keeping most of the bookkeeping out of
the optimizers, while allowing optimizers to be called with problems
structured in different ways.

(The optimizers are still "dense": it still constructs a full Jacobian
(i.e., not sparse), and passes that mutably to the `Problem`. With good
decomposition, systems should generally not get so large for that to
become intractable.)

This refactors the way the expressions are evaluated, by making use of
the `VariableMap`, removing the manual per-expression bookkeeping.