Using recursive assembly, solve each constraint exactly once (#100)
In the partial order of clusters, once a constraint has been used to
rigidly solve geometry, subsequent clusters can just use the rigid
positions of that geometry and don't need to solve that constraint
again.

This was already partially in, except for the final, remaining
under-rigid subgraph that gets solved at the end.