Find transitive closure of elements affected during assembly (#99)
When solving using a recursive assembly plan, elements we're now solving
for may occur on one or more existing cluster's frontiers. The previous
code correctly solves for those cluster's poses such that all clusters
agree on those elements' global positions.

However, the previous code does not take into account that changing one
of those clusters' poses may move an element that we're not directly
solving for, but which may itself occur on some other cluster's
frontier. That cluster will then also have to be moved, etc.

Therefore, we need to solve for the transitive closure of all elements
affected.

https://github.com/endoli/fiksi/pull/97 passes once rebased on top of
this.