Implement calculating sparse Cholesky structure
This calculates the sparsity structure of the `L^T`-factor for `LL^T =
A^T A` given sparse `A`. The sparsity structure of `L^T` is equivalent
to that of the `R`-factor in the QR-decomposition `QR = A^T A`.

This follows "Computing Row and Column Counts for Sparse QR and LU
Factorization" (2001) by J. R. Gilbert, X. S. Li, E. G. Ng & B. W.
Peyton to calculate the structure of `L^T` in `LL^T = A^T A` without
explictly forming `A^T A`.

This can use a bit more clean up before landing: for example, we may not
need all of the calculations done here all the time. For example, we can
often do without row counts, or do without the row indices, but I wanted
to open the PR today nevertheless.