Add a test exercising solve order (#110)
This is an interesting case for decomposition, as it requires correct
solving order as well as being able to solve against a global coordinate
system.


Defines points p0, p1, p2, p3. Constrains p2 to be coincident on fixed
p3, and constrains both point pairs (p0, p1) and (p1, p2) to have a
given distance. A good solve order would be to solve in topological
order: first solve p2 to be on p3, then solve p1 relative to p2, then
solve p0 relative to p1. Finding such a solve order purely structurally
is not trivial.