Improve exhaustive dense subgraph finder for recursive assembly (#103)
The old routine does a depth-first search, which doesn't actually find
*minimum* dense subgraphs, which a a breadth-first search does find.
Finding small dense subgraphs is important for solving performance. Note
both the DFS and the BFS implementations have exponential
time-complexity, but BFS will in general lead to constructing more
efficient decomposition plans.

https://github.com/endoli/fiksi/pull/97 introduced a polynomial-time
*minimal* dense subgraph finder (not *minimum*). Its correctness is
harder to check, so I'm thinking to keep the exhaustive search too, for
now, mainly for ease of debugging.

In addition to changing to BFS, this now makes use of the sparsity of
the graph (normally, most elements only have a few incident edges) to
reduce the amount of scanning performed, making this more useful in
practice.