Reuse an outer QR decomposition for damped least-squares (#113)
Provide a nice speed-up for big (sub)systems, especially those that
require many iterations to find the right $$\lambda$$ in
Levenberg-Marquardt.

To calculate the damped least squares in Levenberg-Marquardt, we
calculate $$[J; \sqrt{\lambda} I] \delta = [r; 0]$$. For numerical
robustness, we use a QR-decomposition to do that. Currently we decompose
the full $$[J; \sqrt{\lambda}I]$$ every time $$\lambda$$ changes.

Instead, we can decompose $$QR = J$$, and then solve $$[R; 0;
\sqrt{\lambda} I] \delta = [Q^T r; 0]$$ using QR-decomposition, which
will tend to be faster as $$R$$ is already upper-triangular, and so
almost everything under the diagonal is already 0; the only exception
being the diagonal damping matrix $$\sqrt{\lambda} I$$ at the bottom. We
may be able to do a bit better than this still by making use of the
structure of that damping matrix.