Init port of Tim Davis' `COLAMD`
The file `colamd.rs` was generated using c2rust on the
[COLAMD source files](https://github.com/DrTimothyAldenDavis/SuiteSparse/tree/dev/COLAMD),
only manually changing `#[no_mangle]` to `#[unsafe(no_mangle)]` to allow
compiling on edition 2024, overriding
`SuiteSparse_config_printf_func_get` to explicitly crash (the consuming
code is not reachable from this crate's public interface), and
surpressing a bunch of lints.

The functionality can only be called through the wrapping API (in
`lib.rs`).

This provides a baseline from which parts can be replaced with idiomatic
Rust.