Use per-constraint or root mean square threshold in tests (#168)
Previously we tested the sum of squared residuals, but the sum is very
dependent on the number of constraints. Checking the mean square means
we aren't unnecessarily stringent on large systems, and taking the
square root makes the value be in the same range as the residuals
themselves (e.g., if two points are respectively `0.2` and `0.4` off
according to two distance constraints, the root mean square residual is
`~0.32`).