Exit Levenberg-Marquardt based on "function tolerance"
If the relative change in the cost function is very low, we exit. This
is a scale-invariant exit condition, and can be triggered when we have
successfully converged, even if we haven't converged to a root. Before
this, we always continued iterating until either a root or some other
stopping condition was found, which will be slow to exit in case a root
doesn't exist.