colamd: remove array-based `knobs` in favor of `Options` (#169)
The behavior remains the same (in particular, the default value for
`Options` is the same as what used to be set by `colamd_set_defaults` in
case of missing `knobs`), but the internal conversion to and reading
from an array of `f64`s is removed. For clarity, the `Options` struct is
moved to its own module.