Calculate system scale and divide it out of variables and expressions (#174)
For numeric solving, it's nice if the problem is well-conditioned. If
the problem is very badly conditioned, it may not be solvable or will
solve slowly.

For Levenberg-Marquardt, this means the Jacobian should be
well-conditioned. We bring back the total scale of the problem to be on
the order of ~1.0 by calculating the root mean square of all lengths,
and dividing all lengths (in both element variables and constraints) by
the scale before going into numeric solving.
    
This removes the `O(system scale)` effect on length-like residuals (such
as point-point distance), making, e.g., length and angle residuals
(which are `O(1)` in radians) more comparable.

Note this work on the system as a whole. It's still possible for
geometry within a problem to differ in scale (e.g., a very small detail
included in a very large problem).